--- Server/API/login/login.go
@@ -1,9 +1,0 @@
-package login
-
-import (
-	"github.com/labstack/echo/v4"
-)
-
-func LoginController(e *echo.Echo) {
-	e.POST("/login", postLogin)
-}
