+++ Server/API/rootController.go
@@ -2,8 +2,8 @@ package rootController
 
 import (
 	"github.com/labstack/echo/v4"
+	"ishtaloo.io/API/auth"
 	"ishtaloo.io/API/index"
-	"ishtaloo.io/API/login"
 	"ishtaloo.io/API/rooms"
 	"ishtaloo.io/API/users"
 	"ishtaloo.io/API/words"
@@ -13,6 +13,6 @@ func RootController(e *echo.Echo) {
 	users.UsersController(e)
 	index.IndexController(e)
 	words.WordsController(e)
-	login.LoginController(e)
+	auth.AuthController(e)
 	rooms.RoomsController(e)
 }
